LayoutTests/imported/w3c/web-platform-tests/cors/credentials-flag-expected.txt

CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=True due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=TRUE due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=%22true%22 due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=1 due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=false due to access control checks.
CONSOLE MESSAGE: Credentials flag is true, but Access-Control-Allow-Credentials is not "true".
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?credentials=0 due to access control checks.
CORS - Access-Control-Allow-Credentials


PASS Setting withCredentials on a sync XHR object should not throw 
PASS Don't send cookie by default 
FAIL Don't send cookie part 2 assert_equals: Cookie sent in withCredentials=true sync request expected "COOKIE" but got "NO_COOKIE"
FAIL Don't obey Set-Cookie when withCredentials=false assert_equals: third expected "COOKIE" but got "NO_COOKIE"
PASS Access-Control-Allow-Credentials: TRUE should be disallowed (async) 
PASS Access-Control-Allow-Credentials: True should be disallowed (async) 
PASS Access-Control-Allow-Credentials: "true" should be disallowed (async) 
PASS Access-Control-Allow-Credentials: false should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 1 should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 0 should be disallowed (async) 
